reordered win/api and added more privilege functions

+++ win/api/constants.go
@@ -6,7 +6,7 @@ const (
 
 	// source: https://golang.org/src/internal/syscall/windows/security_windows.go
 	SE_DEBUG_NAME        = "SeDebugPrivilege"
-	SE_PRIVILEGE_ENABLED = 0x00000002
+	SE_PRIVILEGE_ENABLED = uint32(0x00000002)
 
 	// PROCESS_QUERY_INFORMATION|PROCESS_VM_READ
 )
